move pumper into its own file and improve the return function
